+++ src/bq_common/platform/thread/thread.cpp
@@ -14,7 +14,7 @@ bool bq::platform::thread::cancel()
 {
     enum_thread_status running_status = enum_thread_status::running;
     if (!status_.compare_exchange_strong(running_status, enum_thread_status::pendding_cancel)) {
-        bq::util::log_device_console(log_level::warning, "you're trying to cancel a thread which is not in running status, status enum:%d", status_.load());
+        bq::util::log_device_console(log_level::warning, "you're trying to cancel a thread which is not in running status, status enum:%" PRId32 "", static_cast<int32_t>(status_.load()));
         return false;
     }
     return true;
